rename: 연쇄 이동
