+++ src/test/java/io/woori/account/wooriaccount/repository/jpa/AccountRepositoryTest.java
@@ -2,9 +2,8 @@ package io.woori.account.wooriaccount.repository.jpa;
 
 import io.woori.account.wooriaccount.domain.entity.Account;
 import io.woori.account.wooriaccount.domain.entity.Customer;
-import io.woori.account.wooriaccount.dto.user.SignUpRequestDTO;
+import io.woori.account.wooriaccount.dto.customer.SignUpRequestDTO;
 import io.woori.account.wooriaccount.dummy.DummyCustomer;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -19,7 +18,6 @@ import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
 
 
 @DataJpaTest
