+++ src/test/java/io/woori/account/wooriaccount/repository/jpa/TxHistoryRepositoryTest.java
@@ -4,7 +4,7 @@ import io.woori.account.wooriaccount.domain.entity.Account;
 import io.woori.account.wooriaccount.domain.entity.Customer;
 import io.woori.account.wooriaccount.domain.entity.DepositTxHistory;
 import io.woori.account.wooriaccount.domain.entity.WithdrawTxHistory;
-import io.woori.account.wooriaccount.dto.user.SignUpRequestDTO;
+import io.woori.account.wooriaccount.dto.customer.SignUpRequestDTO;
 import io.woori.account.wooriaccount.dummy.DummyCustomer;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -13,13 +13,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 
 import java.math.BigDecimal;
-import java.util.Optional;
 
-import static org.junit.jupiter.api.Assertions.*;
 @DataJpaTest
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
 class TxHistoryRepositoryTest {
